fix(paper_trader): default replay direction to "long" (D29 trailing edge)
Mirrors the JSONL-replay code path to commit 2b17958: positions resurrected
from `paper_positions_*.jsonl` on PaperTrader startup were defaulting to
direction="short", which would re-introduce the D29 direction bug for any
legacy file lacking an explicit `direction` field. Switching the replay
default to "long" matches the live-open path (engine feed is always
negative-funding assets, which we LONG to collect funding).

Per ats_files/decision_log.md D29-D30. Tests: 310/310 passing locally.

+++ src/simulator/paper_trader.py
@@ -186,7 +186,7 @@ class PaperTrader:
                     notional_usd=float(rec.get("notional_usd", 0.0)),
                     entry_funding_apy=float(rec.get("entry_funding_apy", 0.0)),
                     entry_price=float(rec.get("entry_price", 0.0)),
-                    direction=rec.get("direction", "short"),
+                    direction=rec.get("direction", "long"),
                     peak_roe=float(rec.get("peak_roe", 0.0)),
                     current_roe=float(rec.get("current_roe", 0.0)),
                     price_pnl_usd=float(rec.get("price_pnl_usd", 0.0)),
